+++ weblate/accounts/avatar.py
@@ -85,6 +85,28 @@ def avatar_for_email(email, size=80):
     return url
 
 
+def get_fallback_avatar_url(size):
+    """
+    Returns URL of fallback avatar.
+    """
+    return os.path.join(
+        settings.MEDIA_URL,
+        'weblate-{0}.png'.format(size)
+    )
+
+
+def get_fallback_avatar(size):
+    """
+    Returns fallback avatar.
+    """
+    fallback = os.path.join(
+        appsettings.WEB_ROOT,
+        'media/weblate-{0}.png'.format(size)
+    )
+    with open(fallback, 'r') as handle:
+        return handle.read()
+
+
 def get_avatar_image(user, size):
     """
     Returns avatar image from cache (if available) or downloads it.
@@ -112,12 +134,7 @@ def get_avatar_image(user, size):
                 user.username,
                 str(error)
             )
-            fallback = os.path.join(
-                appsettings.WEB_ROOT,
-                'media/weblate-{0}.png'.format(size)
-            )
-            with open(fallback, 'r') as handle:
-                return handle.read()
+            return get_fallback_avatar(size)
 
     return image
 
@@ -159,11 +176,8 @@ def get_user_display(user, icon=True, link=False):
 
     # Icon requested?
     if icon and appsettings.ENABLE_AVATARS:
-        if user is None:
-            avatar = os.path.join(
-                settings.MEDIA_URL,
-                'weblate-32.png'
-            )
+        if user is None or user.email == 'noreply@weblate.org':
+            avatar = get_fallback_avatar_url(32)
         else:
             avatar = reverse(
                 'user_avatar', kwargs={'user': user.username, 'size': 32}