+++ weblate/trans/models/source.py
@@ -19,17 +19,32 @@
 #
 
 from django.db import models
+from django.utils.translation import ugettext_lazy as _
 
 from weblate.trans.models.subproject import SubProject
+
+PRIORITY_CHOICES = (
+    (60, _('Very high')),
+    (80, _('High')),
+    (100, _('Medium')),
+    (120, _('Low')),
+    (140, _('Very low')),
+)
 
 
 class Source(models.Model):
     checksum = models.CharField(max_length=40)
     subproject = models.ForeignKey(SubProject)
     timestamp = models.DateTimeField(auto_now_add=True)
-    priority = models.IntegerField(default=100)
+    priority = models.IntegerField(
+        default=100,
+        choices=PRIORITY_CHOICES,
+    )
 
     class Meta(object):
+        permissions = (
+            ('edit_priority', "Can edit priority"),
+        )
         app_label = 'trans'
         unique_together = ('checksum', 'subproject')
 
@@ -51,3 +66,10 @@ class Source(models.Model):
             old = Source.objects.get(pk=self.pk)
             self.priority_modified = (old.priority != self.priority)
         super(Source, self).save(force_insert, **kwargs)
+
+    @models.permalink
+    def get_absolute_url(self):
+        return ('review_source', (), {
+            'project': self.subproject.project.slug,
+            'subproject': self.subproject.slug,
+        })