+++ docs/changes.rst
@@ -1,6 +1,15 @@
 Changes
 =======
 
+weblate 1.10
+-----------
+
+Released on ? 2014.
+
+* Documentation improvements.
+* Added whiteboard for site wide messages.
+* Configurable strings priority.
+
 weblate 1.9
 -----------